CustomerGroupDraft to interface. #942
